server/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# LangChain & Ecosystem
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters

# Vectorstore
pinecone

# Embeddings
sentence-transformers
langchain-google-generativeai

# PDF Parsing
PyPDF  # backend for PyPDFLoader via langchain_community

# Environment Variables
python-dotenv

# Typing & Utilities
pydantic
requests

# Logging (optional but recommended)
loguru

# File Uploads
aiofiles

# Process
tqdm